Add first tws connection
This connection is based on client version 66 and uses the startapi as mentioned in API v9.72+. This connection is the basic connection option as this will limit the server version to version 76. A lot of new API functionality works with higher server versions.

Two files with all the needed id's have been added. The objects will be hidden from the users.

Print method added for the tws connection.
First documentation added for the print method and the main package documentation.

Description and namespace adjusted for the coming use of glue in the package.