Added is_ connected objects and checks
changed readBin in tws_connect to read 2 lines instead of twice 1 line. In python there is a check if there are more lines than 2, but not in java. So for now leave it at 2.

Created internal function is_tws_connection_open to check if connection is open. Used this in several other functions as a check.

Added tick type object that is used to check the ticks sent by tws to the api.