Fix mistake in internal function
There was a mistake in the internal function that checks if the connection is open. Called the environment instead of the connection.

Adjusted the disconnect to set everything to NULL, 0 or -1 as in python / c++ api.

Set all server versions to numeric. This is later needed for comparing server version.