Package can now connect correctly to tws.
The connecting process to tws is now in a first stage. The handshake can be done and in the connection object is the nextvalidId available. This is called via the reqIds function. A successful connection also sends a managed accounts message. This process still needs to be incorporated in the connection setup.

For now a basic process_message function and an eWrapper function have been created. For now the process_message function will be treated as an internal function. Not sure if the eWrapper should also be moved to internal status. Both functions are called inside other functions for their side-effects.

The reqIds function calls for the next valid id and when connecting is being triggered.

A make_field encoder has been added to the utils.R file to encode fields before being send to tws.

In summary a connection can be established,  checked and disconnected.