Create better tws_connect
Add a timer to tws_connect to wait a bit, then fires of a request for managed accounts and next valid id. Also added the function req_managed_accounts and added all the necessary extension in process_messages and ewrapper functions.

Add a counter in utils.R to use in the ewrapper functions.

Fix a mistake in the is_tws_connection_open function.