Refactor process_messages and remove ewrapper
ewrapper has been removed. It made no sense to keep this wrapper object in the calls when each processing of the messages is handles by the respective calls.

Process_message has been adjusted to need the tws_con object and all the process calls are made to the separate process functions. These functions are internal functions with each function. req_contract_details has 2 processing messages 1 for the contract details and 1 for processing the end message.

All this make testing a bit easier as a mistake can now only be in 2 places instead of 3. Also this prepares everything for mock testing as the functions can now all be called seperately instead of first needing to invoke a wrapper object.

This did mean that all functions needed to be adjusted.

Added req_contract_details function that can call the detailed information with a valid contract and handles the message processing.

Updated several documentation pages with information from the IB website.